# backend/src/config/mobileOpenApi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lidifin Mobile API",
    "version": "1.0.0",
    "description": "Dedicated mobile-facing API contract for native clients. All paths are relative to /api. Primary authentication is username/password via JWT access + refresh tokens."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Relative API base path"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Primary mobile authentication. Obtain with POST /auth/login and refresh with POST /auth/refresh."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Secondary device authentication for paired devices and integrations."
      },
      "streamTokenAuth": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "Optional query token for stream URLs when attaching Authorization headers is inconvenient."
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Web UI session authentication. Included for completeness."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" },
          "details": {}
        }
      },
...
      "TrackArtist": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" }
        }
      },
      "TrackAlbum": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "coverArt": { "type": "string", "nullable": true },
          "artist": { "$ref": "#/components/schemas/TrackArtist" }
        }
      },
      "TrackSummary": {
        "type": "object",
        "required": ["id", "title", "duration"],
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "duration": { "type": "integer" },
          "trackNo": { "type": "integer" },
          "album": { "$ref": "#/components/schemas/TrackAlbum" }
        }
      },
...
    }
  },
  "paths": {
...
    "/library/tracks": {
      "get": {
        "tags": ["Library"],
        "summary": "List library tracks",
        "security": [
          { "bearerAuth": [] },
          { "apiKeyAuth": [] },
          { "sessionAuth": [] }
        ],
        "parameters": [
          { "in": "query", "name": "albumId", "schema": { "type": "string" } },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "default": 100 } },
          { "in": "query", "name": "offset", "schema": { "type": "integer", "default": 0 } },
          { "in": "query", "name": "sortBy", "schema": { "type": "string", "default": "name" } }
        ],
        "responses": {
          "200": {
            "description": "Tracks returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tracks": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/TrackSummary" }
                    },
                    "total": { "type": "integer" },
                    "offset": { "type": "integer" },
                    "limit": { "type": "integer" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "503": {
            "description": "Upstream Jellyfin unavailable",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  }
}